+++ packages/zowe-explorer/src/dataset/DatasetTree.ts
@@ -19,7 +19,7 @@ import { ValidProfileEnum, IZoweTree, IZoweDatasetTreeNode, PersistenceSchemaEnu
 import { Profiles } from "../Profiles";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { FilterDescriptor, FilterItem, resolveQuickPickHelper, errorHandling } from "../utils/ProfilesUtils";
-import { sortTreeItems, getAppName, getDocumentFilePath, labelRefresh, refreshTree } from "../shared/utils";
+import { sortTreeItems, getAppName, getDocumentFilePath, labelRefresh } from "../shared/utils";
 import { ZoweTreeProvider } from "../abstract/ZoweTreeProvider";
 import { ZoweDatasetNode } from "./ZoweDatasetNode";
 import { getIconById, getIconByNode, IconId, IIconItem } from "../generators/icons";
@@ -29,6 +29,7 @@ import { resetValidationSettings } from "../shared/actions";
 import { closeOpenedTextFile } from "../utils/workspace";
 import { PersistentFilters } from "../PersistentFilters";
 import { IDataSet, IListOptions } from "@zowe/cli";
+import { refreshTree } from "../../src/utils/ProfilesUtils";
 
 // Set up localization
 nls.config({
@@ -99,7 +100,10 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
      */
     public async rename(node: IZoweDatasetTreeNode) {
         await Profiles.getInstance().checkCurrentProfile(node.getProfile());
-        if (Profiles.getInstance().validProfile !== ValidProfileEnum.INVALID) {
+        if (
+            Profiles.getInstance().validProfile === ValidProfileEnum.VALID ||
+            Profiles.getInstance().validProfile === ValidProfileEnum.UNVERIFIED
+        ) {
             return contextually.isDsMember(node) ? this.renameDataSetMember(node) : this.renameDataSet(node);
         }
     }
@@ -350,9 +354,25 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
             // If no profile/session yet, then add session and profile to parent profile node in this.mFavorites array:
             try {
                 profile = Profiles.getInstance().loadNamedProfile(profileName);
-                session = ZoweExplorerApiRegister.getMvsApi(profile).getSession();
-                parentNode.setProfileToChoice(profile);
-                parentNode.setSessionToChoice(session);
+                await Profiles.getInstance().checkCurrentProfile(profile);
+                if (
+                    Profiles.getInstance().validProfile === ValidProfileEnum.VALID ||
+                    Profiles.getInstance().validProfile === ValidProfileEnum.UNVERIFIED
+                ) {
+                    session = ZoweExplorerApiRegister.getMvsApi(profile).getSession();
+                    parentNode.setProfileToChoice(profile);
+                    parentNode.setSessionToChoice(session);
+                } else {
+                    return [
+                        new ZoweDatasetNode(
+                            localize("loadProfilesForFavorites.authFailed", "You must authenticate to view favorites."),
+                            vscode.TreeItemCollapsibleState.None,
+                            parentNode,
+                            null,
+                            globals.INFORMATION_CONTEXT
+                        ),
+                    ];
+                }
             } catch (error) {
                 const errMessage: string =
                     localize(
@@ -611,10 +631,11 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
         );
         // Remove profile node from Favorites if it contains no more favorites.
         if (profileNodeInFavorites.children.length < 1) {
-            this.removeFavProfile(profileName, false);
+            return this.removeFavProfile(profileName, false);
         }
         await this.updateFavorites();
         this.refreshElement(this.mFavoriteSession);
+        return;
     }
 
     /**
@@ -658,6 +679,7 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 ignoreFocusOut: true,
                 canPickMany: false,
             };
+            // If user did not select "Continue", do nothing.
             if (
                 (await vscode.window.showQuickPick([continueRemove, cancelRemove], quickPickOptions)) !== continueRemove
             ) {
@@ -665,6 +687,7 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
             }
         }
 
+        // Remove favorited profile from UI
         this.mFavorites.forEach((favProfileNode) => {
             const favProfileLabel = favProfileNode.label.trim();
             if (favProfileLabel === profileName) {
@@ -673,6 +696,10 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 this.refresh();
             }
         });
+
+        // Update the favorites in settings file
+        await this.updateFavorites();
+        return;
     }
 
     public async onDidChangeConfiguration(e) {
@@ -833,9 +860,14 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
         this.log.debug(localize("enterPattern.log.debug.prompt", "Prompting the user for a data set pattern"));
         let pattern: string;
         await this.checkCurrentProfile(node);
+        let nonFaveNode;
 
-        if (Profiles.getInstance().validProfile !== ValidProfileEnum.INVALID) {
+        if (
+            Profiles.getInstance().validProfile === ValidProfileEnum.VALID ||
+            Profiles.getInstance().validProfile === ValidProfileEnum.UNVERIFIED
+        ) {
             if (contextually.isSessionNotFav(node)) {
+                nonFaveNode = node;
                 if (this.mHistory.getSearchHistory().length > 0) {
                     const createPick = new FilterDescriptor(DatasetTree.defaultDialogText);
                     const items: vscode.QuickPickItem[] = this.mHistory
@@ -897,12 +929,11 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 pattern = node.getLabel();
                 const sessionName = node.getProfileName();
                 await this.addSession(sessionName);
-                const faveNode = node;
-                node = this.mSessionNodes.find((tempNode) => tempNode.label.trim() === sessionName);
-                if (!node.getSession().ISession.user || !node.getSession().ISession.password) {
-                    node.getSession().ISession.user = faveNode.getSession().ISession.user;
-                    node.getSession().ISession.password = faveNode.getSession().ISession.password;
-                    node.getSession().ISession.base64EncodedAuth = faveNode.getSession().ISession.base64EncodedAuth;
+                nonFaveNode = this.mSessionNodes.find((tempNode) => tempNode.label.trim() === sessionName);
+                if (!nonFaveNode.getSession().ISession.user || !nonFaveNode.getSession().ISession.password) {
+                    nonFaveNode.getSession().ISession.user = node.getSession().ISession.user;
+                    nonFaveNode.getSession().ISession.password = node.getSession().ISession.password;
+                    nonFaveNode.getSession().ISession.base64EncodedAuth = node.getSession().ISession.base64EncodedAuth;
                 }
             }
             // looking for members in pattern
@@ -929,8 +960,9 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 }
                 dsSets.push(dataSet);
             }
-            node.label = node.label.trim() + " ";
-            node.label.trim();
+
+            nonFaveNode.label = nonFaveNode.label.trim() + " ";
+            nonFaveNode.label.trim();
             let datasets: string;
             for (const item of dsSets) {
                 if (item.dataSetName) {
@@ -942,11 +974,11 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 }
             }
             if (datasets) {
-                node.tooltip = node.pattern = datasets.toUpperCase();
+                nonFaveNode.tooltip = nonFaveNode.pattern = datasets.toUpperCase();
             } else {
-                node.tooltip = node.pattern = pattern.toUpperCase();
+                nonFaveNode.tooltip = nonFaveNode.pattern = pattern.toUpperCase();
             }
-            const response = await this.getChildren(node);
+            const response = await this.getChildren(nonFaveNode);
             for (const child of response) {
                 let resetIcon: IIconItem;
                 if (child.collapsibleState === vscode.TreeItemCollapsibleState.Collapsed) {
@@ -958,6 +990,7 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                 if (resetIcon) {
                     child.iconPath = resetIcon.path;
                 }
+
                 // remove any previous search memberPatterns
                 if (child.contextValue.includes(globals.FILTER_SEARCH)) {
                     child.contextValue = child.contextValue.replace(globals.FILTER_SEARCH, "");
@@ -1016,11 +1049,11 @@ export class DatasetTree extends ZoweTreeProvider implements IZoweTree<IZoweData
                         }
                     }
                 }
-                node.collapsibleState = vscode.TreeItemCollapsibleState.Expanded;
-                node.dirty = true;
-                const icon = getIconByNode(node);
+                nonFaveNode.collapsibleState = vscode.TreeItemCollapsibleState.Expanded;
+                nonFaveNode.dirty = true;
+                const icon = getIconByNode(nonFaveNode);
                 if (icon) {
-                    node.iconPath = icon.path;
+                    nonFaveNode.iconPath = icon.path;
                 }
             }
             this.addSearchHistory(pattern);